src/requirements.txt
black==v19.10b0
flake8>=3.7.9, <4.0
ipython~=7.0
isort>=4.3.21, <5.0
kedro[pandas.CSVDataSet]==0.17.0
nbstripout==0.3.3
pytest-cov~=2.5
pytest-mock>=1.7.1, <2.0
pytest~=5.0
wheel==0.32.2
psycopg2==2.8.5
SQLAlchemy==1.3.16
typing==3.7.4
python-dateutil==2.8.1
python-crontab==2.5.1